+++ src/db/models/chunk.rs
@@ -6,7 +6,7 @@ pub struct Chunk {
     pub index: i64,
     pub start_block_number: i64,
     pub end_block_number: i64,
-    pub total_tx_num: i64,
+    pub total_tx_num: i32,
     pub hash: String,
     pub batch_hash: Option<String>,
     pub created_at: NaiveDateTime,
